+++ src/Services/PropertyService/PropertyService.Domain/Entities/Image.cs
@@ -10,7 +10,7 @@ public class Image : AuditableEntity
     public required string BucketName { get; set; }
     public required string ObjectName { get; set; }
     public required string ContentType { get; set; }
-    public required decimal FileSize { get; set; }
+    public required double FileSize { get; set; }
     public bool IsMain { get; set; }
     public Guid? PropertyId { get; set; }
     public Guid? ProjectId { get; set; }
